pyproject.toml
 [build-system]
 requires = ["setuptools>=68", "wheel"]
 build-backend = "setuptools.build_meta"

 [project]
 name = "ai-sub-monitor"
 version = "0.1.0"
 description = "CLI to monitor pricing, rate limits, and sentiment for leading AI subscription products."
 authors = [{ name = "Your Name" }]
 requires-python = ">=3.11"
 dependencies = [
   "click>=8.1.7",
   "httpx>=0.27.0",
   "beautifulsoup4>=4.12.3",
   "pyyaml>=6.0.1",
   "sqlalchemy>=2.0.25",
   "aiosqlite>=0.20.0",
   "jinja2>=3.1.3",
   "rich>=13.7.0",
   "openpyxl>=3.1.2",
   "praw>=7.7.0",
   "PyGithub>=2.3.0",
   "vaderSentiment>=3.3.2",
 ]

 [project.optional-dependencies]
 dev = [
   "pytest>=7.4.4",
   "pytest-asyncio>=0.23.3",
   "ruff>=0.1.14",
   "mypy>=1.8.0",
 ]

 [project.scripts]
 ai-sub-monitor = "ai_sub_monitor.cli:cli"

 [tool.setuptools.packages.find]
 where = ["src"]
